Started on the structure of the village package
Want to start up on the way it will function before i forgot. Basicly
how this works is the village act as part of an empire, and at the same
time as its own entity. It will function along the same lines as a tile
entity. As well a village will have sub objects of both villagers and
buildings. Building will act like tile entities and work with the
village. They are mainly just used to describe what the building is.
They as well help the NPCs understand what is going on.